saithrift: tests: Fix L2 tests passing considering bridge model
Fixed the following cases in sail2.py tests:

    1) FDB/Vlan/STP creation automatically obtains bridge port by
        specified logical port

    2) LAG creation automatically creates bridge port, which is needed
        for FDB/VLAN member creation.

    3) LAG member creation automatically removes the specified port from
        the default bridge.

    4) LAG member removing automatically creates back port into the default bridge.

All aboves cases are handled in the helpers in the switch.py module with
minimal fixing of sail2.py:L2LagTest. Currently the default bridge is
obtained on the switch init flow and it is used in the aboves helpers.
These helpers might be extended with optional parameters to specify
exact bridge id, or flag which tells if it is needed ti create/remove
bridge port automatically (in case of lag & lag member helpers).

Signed-off-by: Vadim Kochan <vadymk@mellanox.com>