Change SAI_MIRROR_SESSION_ATTR_VLAN_HEADER_VALID to CREATE_AND_SET
This attribute can be modified because the destination could be
updated to either within a VLAN or not.

Because SAI_MIRROR_SESSION_ATTR_VLAN_ID is no longer set as
mandatory on create, user needs to set this VLAN ID specifically
when creating an RSPAN session. Otherwise, the default value 0
will be used.

Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>