Set the analyzer port as invalid when SPAN type is remote (#867)
For both local and enhanced remote SPAN the mirrored packet is always unicasted (i.e., sent via a single port). Remote or enhanced remote SPAN in NPU perspective, is just adding the appropriate encapsulation (VLAN or GRE tunnel) to the data packet and sent out via the analyzer port.
But the RSPAN implementation by most of Network OS, normally assigns a special VLAN in which source learning is disabled and the packets to the analyzer is always flooded within the VLAN (i.e., Tx to all VLAN members). In order to achieve this behavior we now have to create multiple mirroring sessions, one for each member in the VLAN by adding the VLAN member as the analyzer port.
Instead we can consider the VLAN ID parameter alone and enable the packet to be mirrored to all its members and thereby making the analyzer port as invalid for remote SPAN.