Add support for Aggregate reporting for TAM Events (#991)
Add an optional attribute to tam/report object to add support for Aggregate reporting for TAM Events. Since it is an optional attribute, the current functionality of reporting every event is unaffected. User can select 'aggregate' mode for the report mode attribute and specify an interval. When chosen, aggregate-reports are sent at the chosen interval.