Add functions for packet allocate, free Summary: Sending a packet requires allocating it first. Typically, this requires allocating from a special pool of DMAed memory. For this a ASIC vendor SDK may support custom allocate and free functions. So mirror those in the SAI spec.
Without such a ability we are left to allocate heap memory,
pass it to send_hostif_packet_fn, which then allocates from the
afore mentioned DMA memory, copies the contents of heap memory
and sends the packet along. This causes us to have a extra
allocation and copy. This can induce a substantial slowdown -
~30% in our measurements

Test Plan: build

Signed-off-by: Jasmeet Bagga <jasmeetbagga@fb.com>