Add packet allocate, free function to allow for 0 copy packet TX (#1137)
* Add functions for packet allocate, free
Summary:
Sending a packet requires allocating it first. Typically, this requires allocating from a special pool of DMAed memory.
For this a ASIC vendor SDK may support custom allocate and free functions. So mirror those in the SAI spec.

Without such a ability we are left to allocate heap memory, pass it to send_hostif_packet_fn, which then allocates from the
afore mentioned DMA memory, copies the contents of heap memory and sends the packet along. This causes us to have a extra allocation and copy. This can induce a substantial slowdown - ~30% in our measurements

* Remove attr parameters from host if packet free functions
Summary:
Generally no attributes need to be consulted for freeing packets, so remove those from the free api. As and when such a capability is needed it can be added as another extension

* Add a attribute to convey that we are doing zero copy TX
Summary:
As titled. Add SAI_HOSTIF_PACKET_ATTR_ZERO_COPY_TX, packet send functions can use this attribute to skip
a extra allocation + copy cycles. For legacy implementations nothing changes, since they can just continue to
do a second allocation and copy from user provided buffer

* Attribute to query packet DMA pool size
Summary:
This can be used in conjuction with sai_allocate_hostif_packet_fn to manage a pool of memory in application vs going to the driver before each packet transfer

Signed-off-by: Jasmeet Bagga <jasmeetbagga@fb.com>