Do not call sai_metadata_sai get APIs before checking if they are allocated (#1182)
New asic - fabric - doesn't have attributes like VLAN, HOSTIF... So Broadcom SAI will not allocate these APIs or function pointers of those API objects are not allocated. This causes  attempts that try to get/read these attributes to fail. So sai discovery for fabric will crash.

Solution: Modify `parse.pl` to verify sai_metadata_sai if they are allocated or not before trying to  attributes.

This PR will change only for getting attributes. Create, Set, Remove may also need to be changed, but this is out of the scope of this PR.

Signed-off-by: ngocdo <ngocdo@arista.com>
Signed-off-by: Jasmeet Bagga <jasmeetbagga@fb.com>