New attribute to start fw broadcast process
We need additional attribute to support 5-step FW broadcast process. This attribute is only applicable to PHY and its valid when broadcast is enabled. This is mainly used when the design has multiple PHYs and user would like to take advantage of broadcast FW download to save MDIO cycles and time.

Signed-off-by: Sid Ukidve <sidharaj-u.ukidve@broadcom.com>