Added new port type for Recycle Port (#1204)
For #1203

It is helpful in certain scenarios for packets that have gone through a full switch pipeline to be re-injected into the beginning (Ingress) of the pipeline. SONiC on a disaggregated VOQ chassis uses this capability. There are other examples. Currently this can only be done via SAI by putting a port in loopback configuration. But this takes a port away from regular network traffic. What is needed is a dedicated Recycle port. Such a port must be treated like a regular network port for all ASIC programming purposes (routing, bridging, ACL, queuing, tunnels, mirroring …). The only exception would be that it does not need to support some MAC layer functions.

Signed-off-by: SarathBug sarath.gadagottu@broadcom.com
Signed-off-by: Sid Ukidve <sidharaj-u.ukidve@broadcom.com>