Add attribute to set FEC and port speed.
There are cases where we technically cannot change from one port
speed/FEC combination to another using the existing single-attribute
write.

As an example: 40G/FC-FEC to 100G/RS-FEC - it is not valid to change to
100G with FC-FEC enabled, and likewise it is not valid to enable RS-FEC
on a 40G link.

Solve this by defining a new struct to handle setting FEC and speed at
the same time.

Signed-off-by: Mike Beresford <mberes@google.com>