Add label attribute for counter (#1407)
Counter does not have any mandatory attribute to identify object uniquely. Adding a label attribute that can be used to uniquely identify object during warmboot scenarios. This is similar to existing label attribute used by LAG and virtual router.
The attribute is considered as user data attached to the object.

Related to:
#1158

Signed-off-by: Midhun Somasundaran msomasundaran@fb.com