[SAI-PTF]Align the sai rpc front name with swss and syncd
In swss and orgagent saiserver front end has a method named as start_sai_thrift_rpc_server. but in new sai_ptf structure it renamed as start_p4_sai_thrift_rpc_server.
In order to make the new saiserver integrate with syncd and only cause
the change in sai. change the name in sai_ptf

Signed-off-by: richardyu-ms <richard.yu@microsoft.com>