[sai-ptf]Init port base on available interface and optimize init process  (#1726)
* [sai-ptf]Assign port configurations base on config file
why
Enhancement on #1725
In some situations, some of the ports might not available, but those ports can be listed from devices, and in the test, those tests intented not to be tested. We need some mechanism to skip the un-available port during port init process.
in #1725, we introduced two configuration files, and the number of the port that for testing depends on CLI parameters and the port_config.ini

There are some drawback of this method. Port onfigure process mix with port turn up process.
We cannot configure the port from all the configurations. We need to split the port config and init(turn up) process.

In this change, the actual port for testing only depends on the CLI parameters, interface. For this parameter, it illustrate how many port are availale from PTF running environment, and also define how many PTF interfaces map to DUT devices. If not map the interface, then those port from DUT should be non-available for test.

Base on this, we turn up the port base on the interfaces and init all the interfaces' setting.

How
make the port init process only depends on the configuration file
make the port turn up process depends on interface parameter
Optimize the process when creating bridge port
Verify
Test in DUT
Verified in Pipeline
T0 test set on s6000 and dx010
PTF test set on 7260 and dx010

Signed-off-by: richardyu-ms <richard.yu@microsoft.com>