SAI attributes for path tracing for midpoint node . (#1841)
Path tracing provides a record of the packet path as a sequence of interface ids. In addition, it provides a record of end-to-end delay, per-hop delay, and load on each egress interface along the packet delivery path.

Following draft describes path tracing in detail.
https://datatracker.ietf.org/doc/draft-filsfils-spring-path-tracing

This change adds SAI attributes to configure path tracing on mid point node.