Added saithrift support to return sai_object_id for a given system_port_id and read VOQ counters for system port (#1931)
For VOQ Chassis, the following two API's are added

Reads all the system ports in the switch from SAI and returns the sai_object_id of the system port requested by the sonic-mgmt.
Takes the sai_object_id of the system port and returns all the VOQs for that system port.

Signed-off-by: saksarav <sakthivadivu.saravanaraj@nokia.com>