Adding ENI into aspell. (#1962)
The word ENI is a term which means elastic network interface. It is used in DASH project and its API design, hence adding this word into the aspell checker.

This issue didn't show up before very likely because aspell ignores the code, however, once we use it in comments, it starts to fail.