[meta] Add exception for sai_switch_health_data_t
This will add exception for breaking change SER Data #1982.
Since new field is added to existing structure and it changes it size,
but it is only used in notification, we allow this breaking change.

It still can cause issues, since that struct will have different size on
the stack, and potentially it can read garbage memory or cause crash
when using old headers and new SAI library