Updating saimacsec.h to add macsec_port_id to flow obj (#2003)
In the current implementation (without the proposed changes), there is no relation between secure flow, secure channel and the macsec_port_id/port. MACsec user is not able create a flow or secure channel for a particular core. This is an optional attiribute so this change won't break current macsec applications and the new attribute should only be used when there is a need to have a relation between flow, SC, SA and a port.

Signed-off-by: Sid Ukidve <sidharaj-u.ukidve@broadcom.com>